+++ AdventOfCode.ApiService/Day2/Day2Solver.cs
@@ -4,12 +4,12 @@ namespace AdventOfCode.ApiService.Day2;
 
 public class Day2Solver : IDaySolver
 {
-    public int CalculatePartOne(string input)
+    public long CalculatePartOne(string input)
     {
         return PartOne.CountSafeLines(input);
     }
 
-    public int CalculatePartTwo(string input)
+    public long CalculatePartTwo(string input)
     {
         return PartTwo.CountSafeLines(input);
     }
