+++ AdventOfCode.ApiService/Day4/Day4Solver.cs
@@ -2,12 +2,12 @@ namespace AdventOfCode.ApiService.Day4;
 
 public class Day4Solver : IDaySolver
 {
-    public int CalculatePartOne(string input)
+    public long CalculatePartOne(string input)
     {
         return XmasFinder.Count(input);
     }
 
-    public int CalculatePartTwo(string input)
+    public long CalculatePartTwo(string input)
     {
         return CrossMasFinder.Count(input);
     }
